+++ dev/pkgdown.yml
@@ -3,7 +3,7 @@ pkgdown: 1.6.1
 pkgdown_sha: ~
 articles:
   experdesign: experdesign.html
-last_built: 2021-04-22T14:33Z
+last_built: 2021-04-27T17:09Z
 urls:
   reference: https://experdesign.llrs.dev/reference
   article: https://experdesign.llrs.dev/articles
