### 2021-05-07 #3214 @llrs
Hi @arfon, I know JOSS is in reduced service mode and I am not in any hurry, but do you have any update on the scope of the article? Many thanks for providing these resource to the developers and looking forward to its success.
