### 2021-06-03 #3214 @nhejazi
@lpantano Thanks for pinging me for this review -- it looks like an interesting package and paper; however, I will not be able to provide a timely review at present, due to a confluence of external deadlines. Happy to help out again in the future.
