### 2021-06-10 #3214 @lpantano
Hi @stemangiola, would you be interested on reviewing this R tool? thanks!
