### 2021-06-11 #3358 @whedon
[REVIEW]: experDesign: helping performing experiments on batches
**Submitting author:** @llrs (<a href="http://orcid.org/0000-0001-9747-2570">Lluís Revilla Sancho</a>)
**Repository:** <a href="https://github.com/llrs/experDesign" target ="_blank">https://github.com/llrs/experDesign</a>
**Version:** v0.1.1
