### 2021-07-05 #3358 @stemangiola
Hello @llrs,

I have read the article and done my review. I attach it as a WORD files with comments.

Please also have a look at my wishlist above, for the requirements that have not been ticked.

[10.21105.joss.03358.docx](https://github.com/openjournals/joss-reviews/files/6761190/10.21105.joss.03358.docx)
