### 2021-07-05 llrs/experDesign#32 @llrs
JOSS review
Review for JOSS at this issue: https://github.com/openjournals/joss-reviews/issues/3358

Review 1: unmarked sections

- [x] License: Does the repository contain a plain-text LICENSE file with the contents of an OSI approved software license?

This is on file LICENSE.md and LICENSE on the root directory of the package and named on the DESCRIPTION file.
- [ ] Substantial scholarly effort: Does this submission meet the scope eligibility described in the JOSS guidelines

There are few lines of code (but I won't add lines of code unless they are necessary) and it hasn't been cited but I'm not sure what I can do about the other aspects of [the guide](https://joss.readthedocs.io/en/latest/submitting.html#substantial-scholarly-effort).

After reducing the duplicate text, there will be more space available to include some code examples and comparison as explained.

Not sure if I should document it better on the package or on the article.

- [x] A statement of need: Does the paper have a section titled 'Statement of Need' that clearly states what problems the software is designed to solve and who the target audience is?

Review and add if I missed it, or reword it to make it clearer.

- [ ] Quality of writing: Is the paper well written (i.e., it does not require editing for structure, language, or writing quality)?

I should double check again and probably ask to be reviewed by someone used to check articles.

Comments on the text: [10.21105.joss.03358.docx](https://github.com/llrs/experDesign/files/6762498/10.21105.joss.03358.docx)
Mostly about not precise language, duplicate text, some parts are not clear (mentioning experDesign on previous work is confusing), and no prove of claims
