### 2021-07-06 #3358 @abartlett004
I have also completed my review. I have added my comments to @stemangiola 's Word document and attached it here. I will also add comments regarding my unmarked items to the issue in the repository.

[10.21105.joss.03358_ALB edits.docx](https://github.com/openjournals/joss-reviews/files/6766795/10.21105.joss.03358_ALB.edits.docx)
