### 2021-07-06 llrs/experDesign#32 @abartlett004
Comments for unchecked items on my review checklist:

- Substantial scholarly effort: Does this submission meet the scope eligibility described in the JOSS guidelines

I think this package is useful and I am glad it exists in the world, but it is not a very substantial piece of work. As the JOSS guidelines state, "a minimum allowable contribution should represent not less than three months of work for an individual," which I do not believe is true for this package. I am also not sure what you can do to fix this.

- A statement of need: Does the paper have a section titled 'Statement of Need' that clearly states what problems the software is designed to solve and who the target audience is?

The paper does a good job of explaining what problems the software is designed to solve, but this is described throughout the paper rather than in one specific section. Also, the target audience is not described in much detail.

- Quality of writing: Is the paper well written (i.e., it does not require editing for structure, language, or writing quality)?

There were several instances of unclear or misleading phrasing, improper grammar, and informal terminology. I tried to edit or point out many of these, but I think it would be useful if Joe Moore or another experienced proofreader examined the paper again in detail to fully resolve these issues. Relatedly, I think the language in the introduction section can be made more precise such that it can convey its message in just a few paragraphs, which would leave more space for describing the package and comparing it against existing packages in the description section.
