### 2021-07-06 #3358 @llrs
Many thanks @abartlett004 for your detailed comments and your review! I will start editing the manuscript to reduce duplicated content, add the missing section statement of need, make it more precise and at the same time add more content: code examples, function definition and comparison with other methods.
