### 2021-07-06 llrs/experDesign#32 @llrs
Many thanks @abartlett004 for commenting here too with such great detail and transparency.

About the required substantial scholarly effort; I agree that the package is simple, however it took around 3 months of work to create and simplify the package to the actual shape. However, I might try to add some new features from the open issues to make it worth it.

I will add the  statement of need and describe the target audience. I failed to convey that on the manuscript.

The manuscript will be examined and proofread again after the editions to remove informal terminology and clarify or remove unclear and misleading phrasing. Thanks for the suggestions, I'll do my best.
