+++ paper.Rmd
@@ -1,10 +1,10 @@
 ---
-title: 'experDesign: helping performing experiments on batches'
+title: 'experDesign: stratifyuing samples into batches with minimal bias'
 output:
-  md_document:
-    preserve_yaml: yes
   html_document:
     df_print: paged
+  md_document:
+    preserve_yaml: yes
   word_document:
     keep_md: yes
 date: "23 April 2021"
@@ -30,61 +30,52 @@ tags:
 bibliography: paper.bib
 ---
 
+```{r setup}
+knitr::opts_chunk$set(collapse = TRUE)
+```
+
 # Summary
 
 The design of an experiment is critical to its success.
 Nonetheless, even with a correct design, the process up to the moment of measurement is critical.
 At any one of the several steps required from collection to measurement various errors and problems could affect the experimental results.
 Failure to take such variability into account can render an experiment inconclusive.
-*experDesign* provides tools to minimize the risk of inconclusive results by specifically assigning samples to particular batches.
+*experDesign* provides tools to minimize the risk of inconclusive results by assigning samples to batches to minimize potential batch effects.
 
 # Introduction
 
 To properly design an experiment, the source of the variation between samples must be identified.
 Typically, one can control the environment in which the study or experiment is being conducted.
 Sometimes, however, this is not possible, and then one needs to apply certain techniques to control for such variations.
-There are three techniques used to reduce unwanted variation: blocking, randomization and replication [@klaus2015].
+There are three techniques used to decrease the uncertainty about the unwanted variation: blocking, randomization and replication [@klaus2015].
 
-**Blocking** group samples that are equal, according to one or more variables, allows one to estimate the variation in the measurements of the factors from these variables.
-**Randomization** is a method to minimize the variation in the measurements by mixing the potential confounding variables.
+**Blocking** groups samples that are equal according to one or more variables allowing one to estimate the differences between each batch by comparing measurements within blocks.
+**Randomization** by mixing the potential confounding variables minimizes the variation in the measurements .
 **Replication** increases the number of samples used in an experiment to better estimate the variation of the experiment.
-In some settings (e.g., clinical, agriculture, etc.), several of these techniques can be applied together to ensure the robustness of the study.
+In some settings these techniques can be applied together to ensure the robustness of the study.
 
-Between the designing of an experiment and the measurement of the samples, many things can happen.
-If some samples are lost, are contaminated, or do not pass quality control for measurement purposes, the unwanted variation of the experiment will increase.
-Even if this doesn't happen, experiments will occasionally need to be carried out in batches due to technical reasons; for example, the machine cannot measure more than one samples at a time, or because of other practical considerations; for instance, it may not be possible to obtain additional measurements in the field during the allotted time.
+Between the designing of an experiment and the measurement of the samples, some samples might be lost, be contaminated, or degradation below quality control threshold can happen.
+Even if this does not happen, experiments will occasionally need to be carried out in batches.
+Due to technical, for example, the machine cannot measure more than one sample at a time; and practical reasons, for instance, it may not be possible to obtain additional measurements in the field during the allotted time.
 
+This divergence from the original design might cause batch effects perturbing the analysis.
 There are several techniques to identify and assess batch effects when analyzing an already measured experiment [@leek2010].
-Experiments that run over long periods of time or that are run across different laboratories are highly susceptible to batch effects.
[…]
