+++ paper.Rmd
@@ -1,13 +1,14 @@
 ---
 title: 'experDesign: stratifyuing samples into batches with minimal bias'
 output:
-  html_document:
-    df_print: paged
-  md_document:
-    preserve_yaml: yes
   word_document:
     keep_md: yes
-date: "23 April 2021"
+  md_document:
+    preserve_yaml: yes
+  pdf_document: default
+  html_document:
+    df_print: paged
+date: "15th July 2021"
 affiliations:
 - name: Centro de Investigación Biomédica en Red, Enfermedades Hepáticas y Digestivas
   index: 1
@@ -30,8 +31,8 @@ tags:
 bibliography: paper.bib
 ---
 
-```{r setup}
-knitr::opts_chunk$set(collapse = TRUE)
+```{r setup, include=FALSE}
+knitr::opts_chunk$set(collapse = TRUE, warning = TRUE, error = TRUE)
 ```
 
 # Summary
@@ -69,13 +70,13 @@ Replications consist of increasing the number of measurements with similar attri
 When there is one extraction and then a sample is measured multiple times, this is referred to as a technical replicate.
 Technical replicates help estimate the variation of the measurement method, and thus the possible batch effect.
 
-Randomization and replicates can be utilized to prevent batch effects to confound the analysis.
+Randomization and replicates can be used to prevent batch effects to confound the analysis.
 However, by examining how the variables are distributed across each batch sample, proper randomization can be ensured, thus minimizing batch effects.
 This is known as randomized block experimental design or stratified random sampling experimental design.
 
 # State of the art
 
-There are some tools to prevent batch effects on the R language in multiple fields and areas, and particularly for biological research [@rcoreteam2014].
+There are some tools to minimize batch effects on the R language in multiple fields and areas, and particularly for biological research [@rcoreteam2014].
 Here we briefly describe the currently available packages:
 
 -   *OSAT*, at [Bioconductor](https://bioconductor.org/packages/OSAT/), first allocates the samples from each batch according to a variable; it then shuffles the samples from each batch in order to randomize the other variables [@yan2012].
@@ -173,7 +174,6 @@ There is a bug on Omixer that prevents it to work unless specific conditions are
 
 ```{r Omixer, error=TRUE}
 library("Omixer")
-data(survey, package = "MASS")
 survey$ID <- seq_len(nrow(survey))
 Omixer_index <- omixerRand(survey, sampleId = "ID", 
                            block = "block", iterNum = 10, wells = size_subset, 
@@ -193,9 +193,13 @@ Omixer_index <- omixerRand(survey, sampleId = "ID",
 
 ```
 
-## experDesign
+# Description
 
-***experDesign***, the package described on this paper, first evaluates the distribution of samples in batches and then on a spatial grid.
+The package ***experDesign*** provides the functional `design` to distribute the samples into multiple batches such that each variable is homogeneous within each batch.
+Each batch is set to have some centrality and dispersion statistics to match as closely as possible the original input design data.
+It does by comparing the mean, the standard deviation, the median absolute deviation, variables with no value number, the entropy and the independence of the categorical variables.
+On each iteration if the distribution of samples for each batch has less differences with the original distribution than the last previous stored sample distribution then it replace it as the best sample distribution.
+At the end of the iterations the best sample distribution is returned to the user.
 
 ```{r experDesign}
 library("experDesign")
@@ -220,16 +224,37 @@ experDesign_smoke <- ggplot(experDesign_report) +
 (experDesign_smoke/(experDesign_age + experDesign_sex + plot_layout(guides = 'collect'))) + plot_annotation(title = "experDesign")
 ```
 
-# Description
+One can check the statistics of the index for multiple statistics:
 
[…]
