### 2021-07-20 39feb14
Update paper and add figures
