+++ paper.Rmd
@@ -1,5 +1,5 @@
 ---
-title: 'experDesign: stratifyuing samples into batches with minimal bias'
+title: 'experDesign: stratifying samples into batches with minimal bias'
 output:
   word_document:
     keep_md: yes
@@ -48,30 +48,30 @@ Failure to take such variability into account can render an experiment inconclus
 To properly design an experiment, the source of the variation between samples must be identified.
 Typically, one can control the environment in which the study or experiment is being conducted.
 Sometimes, however, this is not possible, and then one needs to apply certain techniques to control for such variations.
-There are three techniques used to decrease the uncertainty about the unwanted variation: blocking, randomization and replication [@klaus2015].
+There are three techniques used to decrease the uncertainty of the unwanted variation: blocking, randomization and replication [@klaus2015].
 
-**Blocking** groups samples that are equal according to one or more variables allowing one to estimate the differences between each batch by comparing measurements within blocks.
-**Randomization** by mixing the potential confounding variables minimizes the variation in the measurements .
+**Blocking** groups samples that are equal according to one or more variables, allowing one to estimate the differences between each batch by comparing measurements within blocks.
+**Randomization** minimizes the variation in the measurements by mixing the potential confounding variables.
 **Replication** increases the number of samples used in an experiment to better estimate the variation of the experiment.
 In some settings these techniques can be applied together to ensure the robustness of the study.
 
-Between the designing of an experiment and the measurement of the samples, some samples might be lost, be contaminated, or degradation below quality control threshold can happen.
+Between the designing of an experiment and the measurement of the samples, some samples might be lost, be contaminated, or degradation below quality threshold can happen.
 Even if this does not happen, experiments will occasionally need to be carried out in batches.
-Due to technical, for example, the machine cannot measure more than one sample at a time; and practical reasons, for instance, it may not be possible to obtain additional measurements in the field during the allotted time.
+Batches might be due to technical reasons, for example, the machine cannot measure more than some amount of samples at a time; and practical reasons, for instance, it may not be possible to obtain additional measurements in the field during the allotted time.
 
 This divergence from the original design might cause batch effects perturbing the analysis.
 There are several techniques to identify and assess batch effects when analyzing an already measured experiment [@leek2010].
 However, if the source of variations is not taken into account before measurement is completed, a batch effect can be introduced that confounds the analysis.
 Thus, it would be better to avoid such batch effects before executing an experiment.
 
-To prevent the batch effect to confound your analysis after the initial design of the experiment, there are two options: randomization and replication.
+To prevent the batch effect to confound an analysis after the initial design of the experiment, there are two options: randomization and replication.
 Randomization, can help reduce variations across groups, while replication helps estimate the variation of the measurements or samples, thus increasing the precision of the estimates of the true value on the analysis.
 Replications consist of increasing the number of measurements with similar attributes.
-When there is one extraction and then a sample is measured multiple times, this is referred to as a technical replicate.
-Technical replicates help estimate the variation of the measurement method, and thus the possible batch effect.
+When a sample is measured multiple times, this is referred to as a technical replicate.
+Technical replicates help estimate the variation of the measurement method, and thus the possible batch effect [@blainey2014].
 
 Randomization and replicates can be used to prevent batch effects to confound the analysis.
-However, by examining how the variables are distributed across each batch sample, proper randomization can be ensured, thus minimizing batch effects.
+However, by examining how the variables are distributed across each batch, proper randomization can be ensured, thus minimizing batch effects.
[…]
