+++ paper.Rmd
@@ -8,7 +8,7 @@ output:
   pdf_document: default
   html_document:
     df_print: paged
-date: "15th July 2021"
+date: "26st July 2021"
 affiliations:
 - name: Centro de Investigación Biomédica en Red, Enfermedades Hepáticas y Digestivas
   index: 1
@@ -32,7 +32,7 @@ bibliography: paper.bib
 ---
 
 ```{r setup, include=FALSE}
-knitr::opts_chunk$set(collapse = TRUE, warning = TRUE, error = TRUE)
+knitr::opts_chunk$set(collapse = TRUE, warning = TRUE, error = TRUE, fig.width = 9)
 ```
 
 # Summary
@@ -47,17 +47,19 @@ Failure to take such variability into account can render an experiment inconclus
 
 To properly design an experiment, the source of the variation between samples must be identified.
 Typically, one can control the environment in which the study or experiment is being conducted.
-Sometimes, however, this is not possible, and then one needs to apply certain techniques to control for such variations.
+Sometimes, however, this is not possible.
+In this event, techniques to control such variations need to be applied.
 There are three techniques used to decrease the uncertainty of the unwanted variation: blocking, randomization and replication [@klaus2015].
 
-**Blocking** groups samples that are equal according to one or more variables, allowing one to estimate the differences between each batch by comparing measurements within blocks.
-**Randomization** minimizes the variation in the measurements by mixing the potential confounding variables.
+**Blocking** groups samples that are equal according to one or more variables, allowing the estimation of the differences between each batch by comparing measurements within blocks.
+**Randomization** minimizes the variation in the measurements by randomly mixing the potential confounding variables.
 **Replication** increases the number of samples used in an experiment to better estimate the variation of the experiment.
 In some settings these techniques can be applied together to ensure the robustness of the study.
 
-Between the designing of an experiment and the measurement of the samples, some samples might be lost, be contaminated, or degradation below quality threshold can happen.
+Between the designing of an experiment and the measurement of the samples, some samples might be lost, contaminated, or degradated below quality threshold.
 Even if this does not happen, experiments will occasionally need to be carried out in batches.
-Batches might be due to technical reasons, for example, the machine cannot measure more than some amount of samples at a time; and practical reasons, for instance, it may not be possible to obtain additional measurements in the field during the allotted time.
+Batches might be due to technical reasons, for example, the machine cannot measure more than a number of samples at the same time.
+They can be also due to practical reasons, for instance, it may not be possible to obtain additional measurements in the field during the allotted time.
 
 This divergence from the original design might cause batch effects perturbing the analysis.
 There are several techniques to identify and assess batch effects when analyzing an already measured experiment [@leek2010].
@@ -82,18 +84,18 @@ Here we briefly describe the currently available packages:
 -   *OSAT*, at [Bioconductor](https://bioconductor.org/packages/OSAT/), first allocates the samples from each batch according to a variable; it then shuffles the samples from each batch in order to randomize the other variables [@yan2012].
     This algorithm relies on categorical variables and cannot use numerical variables (e.g., age- or time-related) unless they are treated as categorical.
 
--   *minDiff*, at [github](https://github.com/m-Py/minDiff), and its successor *anticlust*, at [CRAN](https://cran.r-project.org/package=anticlust), divides the samples into similar groups, ensuring similarity by enforcing heterogeneity within groups [@papenberg2020].
+-   *anticlust*, at [CRAN](https://cran.r-project.org/package=anticlust), divides the samples into similar groups, ensuring similarity by enforcing heterogeneity within groups [@papenberg2020].
     Conceptually it is similar to the clustering methods k-means.
 
[…]
