+++ paper.Rmd
@@ -45,40 +45,40 @@ Failure to take such variability into account can render an experiment inconclus
 
 # Introduction
 
-To properly design an experiment, the source of the variation between samples must be identified.
+To design an experiment that can support conclusive results upon analysis, the source of the variation between samples must be identified.
 Typically, one can control the environment in which the study or experiment is being conducted.
 Sometimes, however, this is not possible.
 In this event, techniques to control such variations need to be applied.
 There are three techniques used to decrease the uncertainty of the unwanted variation: blocking, randomization and replication [@klaus2015].
 
-**Blocking** groups samples that are equal according to one or more variables, allowing the estimation of the differences between each batch by comparing measurements within blocks.
+**Blocking** is a method that groups samples that are equal according to one or more variables, allowing the estimation of the differences between each batch by comparing measurements within blocks.
 **Randomization** minimizes the variation in the measurements by randomly mixing the potential confounding variables.
 **Replication** increases the number of samples used in an experiment to better estimate the variation of the experiment.
 In some settings these techniques can be applied together to ensure the robustness of the study.
 
 Between the designing of an experiment and the measurement of the samples, some samples might be lost, contaminated, or degradated below quality threshold.
-Even if this does not happen, experiments will occasionally need to be carried out in batches.
-Batches might be due to technical reasons, for example, the machine cannot measure more than a number of samples at the same time.
+In addition, experiments will occasionally need to be carried out in batches.
+Batches might be necessary for technical reasons, for example, the device cannot measure more than a number of samples at the same time.
 They can be also due to practical reasons, for instance, it may not be possible to obtain additional measurements in the field during the allotted time.
 
 This divergence from the original design might cause batch effects perturbing the analysis.
 There are several techniques to identify and assess batch effects when analyzing an already measured experiment [@leek2010].
-However, if the source of variations is not taken into account before measurement is completed, a batch effect can be introduced that confounds the analysis.
-Thus, it would be better to avoid such batch effects before executing an experiment.
+It would be better to avoid such batch effects before executing an experiment.
+By taking into account the de differences between the original design and the state before measurement is done, confounding effects could be minimized.
 
-To prevent the batch effect to confound an analysis after the initial design of the experiment, there are two options: randomization and replication.
-Randomization, can help reduce variations across groups, while replication helps estimate the variation of the measurements or samples, thus increasing the precision of the estimates of the true value on the analysis.
+To prevent the batch effect fom confounding the analysis after the initial design of the experiment, there are two options: randomization and replication.
+Randomization, consists on shuffling the samples to mix samples with different attributes, can help reduce variations across groups, while replication helps estimate the variation of the measurements or samples, thus increasing the precision of the estimates of the true value on the analysis.
 Replications consist of increasing the number of measurements with similar attributes.
 When a sample is measured multiple times, this is referred to as a technical replicate.
 Technical replicates help estimate the variation of the measurement method, and thus the possible batch effect [@blainey2014].
 
 Randomization and replicates can be used to prevent batch effects to confound the analysis.
-However, by examining how the variables are distributed across each batch, proper randomization can be ensured, thus minimizing batch effects.
+By examining how the variables are distributed across each batch, proper randomization can be ensured, thus minimizing batch effects.
 This is known as randomized block experimental design or stratified random sampling experimental design.
 
 # State of the art
 
-There are some tools to minimize batch effects on the R language in multiple fields and areas, and particularly for biological research [@rcoreteam2014].
[…]
