### 2021-08-25 6dee8d8
Following some Joe comments on the text
