+++ paper.Rmd
@@ -8,7 +8,7 @@ output:
   pdf_document: default
   html_document:
     df_print: paged
-date: "26st July 2021"
+date: "25th August 2021"
 affiliations:
 - name: Centro de Investigación Biomédica en Red, Enfermedades Hepáticas y Digestivas
   index: 1
@@ -38,67 +38,68 @@ knitr::opts_chunk$set(collapse = TRUE, warning = TRUE, error = TRUE, fig.width =
 # Summary
 
 The design of an experiment is critical to its success.
-Nonetheless, even with a correct design, the process up to the moment of measurement is critical.
-At any one of the several steps required from collection to measurement various errors and problems could affect the experimental results.
+Nonetheless, even when correctly designed, the process leading up to the moment of measuring a given variable is critical.
+At any one of the several steps, from sample collection to measurement of a variable, various errors and problems can affect the experimental results.
 Failure to take such variability into account can render an experiment inconclusive.
-*experDesign* provides tools to minimize the risk of inconclusive results by assigning samples to batches to minimize potential batch effects.
+*experDesign* provides tools to minimize the risk of inconclusive results by assigning samples to batches to reduce potential batch effects.
 
 # Introduction
 
 To design an experiment that can support conclusive results upon analysis, the source of the variation between samples must be identified.
 Typically, one can control the environment in which the study or experiment is being conducted.
 Sometimes, however, this is not possible.
-In this event, techniques to control such variations need to be applied.
-There are three techniques used to decrease the uncertainty of the unwanted variation: blocking, randomization and replication [@klaus2015].
+In such cases, techniques to control variations must to be applied.
+There are three methods used to decrease the uncertainty of the unwanted variation: blocking, randomization and replication [@klaus2015].
 
-**Blocking** is a method that groups samples that are equal according to one or more variables, allowing the estimation of the differences between each batch by comparing measurements within blocks.
+**Blocking** is a method that groups samples that are equal according to one or more variables, allowing the estimation of the differences between each batch by comparing measurements within the blocks.
 **Randomization** minimizes the variation in the measurements by randomly mixing the potential confounding variables.
 **Replication** increases the number of samples used in an experiment to better estimate the variation of the experiment.
-In some settings these techniques can be applied together to ensure the robustness of the study.
+In some settings these techniques can be applied together to enhance the robustness of the study.
 
-Between the designing of an experiment and the measurement of the samples, some samples might be lost, contaminated, or degradated below quality threshold.
+Between the designing of an experiment and the measurement of the samples, some samples might be lost, contaminated, or degraded below the quality threshold.
 In addition, experiments will occasionally need to be carried out in batches.
-Batches might be necessary for technical reasons, for example, the device cannot measure more than a number of samples at the same time.
-They can be also due to practical reasons, for instance, it may not be possible to obtain additional measurements in the field during the allotted time.
+The later might be needed for technical reasons; for example, the device cannot measure more than a given number of samples at the same time.
+Practical reasons can also be a factor; for instance, it may not be possible to obtain additional measurements in the field during the allotted time.
 
-This divergence from the original design might cause batch effects perturbing the analysis.
+This divergence from the original design might cause batch effects, thereby perturbing the analysis.
 There are several techniques to identify and assess batch effects when analyzing an already measured experiment [@leek2010].
 It would be better to avoid such batch effects before executing an experiment.
-By taking into account the de differences between the original design and the state before measurement is done, confounding effects could be minimized.
[…]
