+++ paper.md
@@ -8,7 +8,7 @@ output:
   pdf_document: default
   html_document:
     df_print: paged
-date: "15th July 2021"
+date: "25th August 2021"
 affiliations:
 - name: Centro de Investigación Biomédica en Red, Enfermedades Hepáticas y Digestivas
   index: 1
@@ -36,65 +36,68 @@ bibliography: paper.bib
 # Summary
 
 The design of an experiment is critical to its success.
-Nonetheless, even with a correct design, the process up to the moment of measurement is critical.
-At any one of the several steps required from collection to measurement various errors and problems could affect the experimental results.
+Nonetheless, even when correctly designed, the process leading up to the moment of measuring a given variable is critical.
+At any one of the several steps, from sample collection to measurement of a variable, various errors and problems can affect the experimental results.
 Failure to take such variability into account can render an experiment inconclusive.
-*experDesign* provides tools to minimize the risk of inconclusive results by assigning samples to batches to minimize potential batch effects.
+*experDesign* provides tools to minimize the risk of inconclusive results by assigning samples to batches to reduce potential batch effects.
 
 # Introduction
 
-To properly design an experiment, the source of the variation between samples must be identified.
+To design an experiment that can support conclusive results upon analysis, the source of the variation between samples must be identified.
 Typically, one can control the environment in which the study or experiment is being conducted.
-Sometimes, however, this is not possible, and then one needs to apply certain techniques to control for such variations.
-There are three techniques used to decrease the uncertainty of the unwanted variation: blocking, randomization and replication [@klaus2015].
+Sometimes, however, this is not possible.
+In such cases, techniques to control variations must to be applied.
+There are three methods used to decrease the uncertainty of the unwanted variation: blocking, randomization and replication [@klaus2015].
 
-**Blocking** groups samples that are equal according to one or more variables, allowing one to estimate the differences between each batch by comparing measurements within blocks.
-**Randomization** minimizes the variation in the measurements by mixing the potential confounding variables.
+**Blocking** is a method that groups samples that are equal according to one or more variables, allowing the estimation of the differences between each batch by comparing measurements within the blocks.
+**Randomization** minimizes the variation in the measurements by randomly mixing the potential confounding variables.
 **Replication** increases the number of samples used in an experiment to better estimate the variation of the experiment.
-In some settings these techniques can be applied together to ensure the robustness of the study.
+In some settings these techniques can be applied together to enhance the robustness of the study.
 
-Between the designing of an experiment and the measurement of the samples, some samples might be lost, be contaminated, or degradation below quality threshold can happen.
-Even if this does not happen, experiments will occasionally need to be carried out in batches.
-Batches might be due to technical reasons, for example, the machine cannot measure more than some amount of samples at a time; and practical reasons, for instance, it may not be possible to obtain additional measurements in the field during the allotted time.
+Between the designing of an experiment and the measurement of the samples, some samples might be lost, contaminated, or degraded below the quality threshold.
+In addition, experiments will occasionally need to be carried out in batches.
+The later might be needed for technical reasons; for example, the device cannot measure more than a given number of samples at the same time.
+Practical reasons can also be a factor; for instance, it may not be possible to obtain additional measurements in the field during the allotted time.
 
[…]
