### 2021-09-01 #3358 @llrs
Yes, sorry for the long silence. I have an updated version of the manuscript almost ready to be updated again. There weren't many (any) issues with the software.

However, I have a doubt about the "Substantial scholarly effort" criteria. It was approved for review with a comment about the small size of the software. Both reviewers didn't mark the "Substantial scholarly effort" (I think because the lines of code are not above the threshold) but I haven't found any clue on which direction the package would need to be expanded or if the manuscript would be approved to publication without more substantial effort on the software.
Many thanks for checking the status.
