### 2021-09-02 6e7b408
Remove long lines moves tatement of need below code examples
