+++ paper.md
@@ -89,13 +89,6 @@ Here we briefly describe the currently available packages:
 -   Recently, [*Omixer*](https://bioconductor.org/packages/Omixer/), a new package, has been made available at Bioconductor [@sinke2021].
     It tests whether the random assignments made by it are homogeneous by transforming all variables to numeric values and using the Kendall's correlation when there are more than 5 samples; otherwise, it utilizes the Pearson's chi-squared test.
 
-# Statement of need
-
-Current solutions for stratifying samples to reduce and control batch effect do not work for all cases.
-They are either specialized to a particular type of data, they omit some conditions that are usually met, or they only work under a specific subset of conditions.
-The new package ***experDesign*** works with all data types and does not require a spatial distribution making it suitable for all kind of experiments.
-This package is intended for people needing a quick and easy solution that will provide reasonable suggestions on how to best distribute the samples for analysis.
-
 For completeness a description and comparison of the usage of the different software packages currently available on CRAN and Bioconductor is presented below.
 First, we start with some real data obtained from a survey:
 
@@ -124,8 +117,10 @@ theme_set(theme_plots)
 library("OSAT")
 library("patchwork")
 gs <- setup.sample(survey, optimal = VoI)
-gc <- setup.container(plate = IlluminaBeadChip96Plate, n = n_batch, batch = 'plates')
-gSetup <- create.optimized.setup(sample = gs, container = gc, nSim = iterations)
+gc <- setup.container(plate = IlluminaBeadChip96Plate, 
+                      n = n_batch, batch = 'plates')
+gSetup <- create.optimized.setup(sample = gs, container = gc, 
+                                 nSim = iterations)
 ## Warning in create.optimized.setup(sample = gs, container = gc, nSim =
 ## iterations): Using default optimization method: optimal.shuffle
 osat_report <- get.experiment.setup(gSetup)
@@ -175,7 +170,8 @@ anticlust_index <- anticlustering(
   method = "exchange",
   repetitions = iterations
 )
-report_anticlust <- data.frame(Age = survey[, "Age"], batch = as.character(anticlust_index))
+report_anticlust <- data.frame(Age = survey[, "Age"], 
+                               batch = as.character(anticlust_index))
 library("ggplot2")
 ggplot(report_anticlust) + 
   geom_histogram(aes(Age, fill = batch)) +
@@ -200,28 +196,40 @@ Omixer_index <- omixerRand(survey, sampleId = "ID",
                            plateNum = n_batch, randVars = VoI, techVars = NULL)
 ## Error in omixerRand(survey, sampleId = "ID", block = "block", iterNum = 10, : Number of unmasked wells must equal number of samples.
 Omixer_index <- omixerRand(survey, sampleId = "ID", 
-                           block = "block", iterNum = 10, wells = size_subset, mask = rep(1, nrow(survey)),
+                           block = "block", iterNum = 10, wells = size_subset, 
+                           mask = rep(1, nrow(survey)),
                            plateNum = n_batch, randVars = VoI, techVars = NULL)
 ## Error: Tibble columns must have compatible sizes.
 ## * Size 288: Existing data.
 ## * Size 237: Column `mask`.
 ## ℹ Only values of size one are recycled.
 Omixer_index <- omixerRand(survey, sampleId = "ID", 
[…]
