### 2021-09-02 #3358 @llrs
@lpantano Submitting this new version for further comments and if decided to publication.

Here are some comments on the review @stemangiola.:
- The license is both on the DESCRIPTION of the package and on the LICENSE.md and LICENSE file on the root directory with an approved OSI license in plain text.
- I have provided examples on the paper about how to use the program. In addition to reviewing the code examples.
- The functions are now also described on the manuscript beside the package own documentation. Hope it is now clear how do they work now.

Both of you @stemangiola and @abartlett004 raised some concerns regarding the scholarly effort, statement of need and quality of writing. See below for my comments:
- Substantial scholarly effort: The amount of work on the software was of 3 months between thinking about the implementation and implementing in few lines of code. If there is anything lacking to the software I would gladly add it to extend and make it more useful. There are some ideas on the issue tracker but I don't think they are important, and they might blur the focus of the package.
- Statement of need: A new section has been added with a statement of need.
- Quality of writing: The manuscript has been through-fully checked and reorganized for better readability. Apologies for all the mistakes it had.
