### 2021-09-09 llrs/experDesign#32 @llrs
I think I fixed all the issues and replied on openjournals/joss-reviews#3358.
If further modifications are needed I'll reopen the issue
