### 2021-09-21 #3358 @lpantano
@stemangiola , @abartlett004, could you give us a timeframe for reviewing the final items that are unchecked in the list above. If something is still missing, you can post the comments here and the authors can address them.
