### 2021-09-21 #3358 @lpantano
@llrs, the one thing that I couldn't find is some kind of score telling you how good or bad is the final design. I can imagine that there are cases where is not ideal, like in one of your example, and it would be good to have some score or warning telling you that even trying to randomize the design, the design will be biased by some degree. Let me know your thoughts. Maybe it is already there and it only needs a better place in the documentation. thanks!
