### 2021-09-21 #3358 @llrs
@lpantano such score is already available on the package. One can explore how good a fit is with `check_index` but this is only between subsets, not among all the combinations. The closer to 0, the better a subset is. I will update the documentation of the function with more comments pointing users to increase the iterations and mention the `check_index` to the vignette too. Thanks for the suggestion.
