### 2021-09-23 #3358 @abartlett004
Thanks for making the revisions! The paper overall looks much better to me now. The writing quality is much improved and the discussion and examples of other existing packages really helped me to understand why your package is substantial. My final comment is that I think you forgot to insert the flowchart regarding which functions to use. Other than that, everything looks good!
