### 2021-09-27 #3358 @llrs
@abartlett004 Apologies, forgot to commit the flowchart (see the new updated version). Many thanks for your encouraging words and review, it has helped to show the value of the package.
