### 2021-11-04 #3358 @lpantano
Hi @llrs, I think the last modification would be related to the length of the paper. The idea is that the papers is 2-3 page long. You could point to all the examples in the paper to the documentation. If there are examples that are not in the docs, you can add it there. Maybe you can do a panel of 2-3 figures of the most important examples and leave the code out (mentioning where you can find that in the docs). Once this is re-formatted, we are good to accept the paper finally. Thank you!
