+++ paper.Rmd
@@ -1,13 +1,13 @@
 ---
 title: 'experDesign: stratifying samples into batches with minimal bias'
 output:
-  word_document:
-    keep_md: yes
-  md_document:
-    preserve_yaml: yes
-  pdf_document: default
   html_document:
     df_print: paged
+  word_document:
+    keep_md: yes
+  pdf_document: default
+  md_document:
+    preserve_yaml: yes
 date: "25th August 2021"
 affiliations:
 - name: Centro de Investigación Biomédica en Red, Enfermedades Hepáticas y Digestivas
@@ -31,8 +31,10 @@ tags:
 bibliography: paper.bib
 ---
 
-```{r setup, include=FALSE}
-knitr::opts_chunk$set(collapse = TRUE, warning = TRUE, error = TRUE, fig.width = 9)
+```{r setup}
+knitr::opts_chunk$set(collapse = TRUE, warning = TRUE, 
+                      error = 0,
+                      fig.width = 9)
 ```
 
 # Summary
@@ -92,9 +94,10 @@ Here we briefly describe the currently available packages:
     It tests whether the random assignments made by it are homogeneous by transforming all variables to numeric values and using the Kendall's correlation when there are more than 5 samples; otherwise, it utilizes the Pearson's chi-squared test.
 
 For completeness a description and comparison of the usage of the different software packages currently available on CRAN and Bioconductor is presented below.
-First, we start with some real data obtained from a survey:
+First, we start with some real data obtained from a survey.
+This data set has three variables of interest; Sex, Smoke and Age are a mix of categorical and numeric variables.
 
-```{r example-data}
+```{r example-data, include = FALSE}
 data(survey, package = "MASS")
 survey$ID <- seq_len(nrow(survey))
 VoI <- c("Sex", "Smoke", "Age")
@@ -103,14 +106,12 @@ size_subset <- 96
 iterations <- 1000
 ```
 
-This data set has three variables of interest; Sex, Smoke and Age are a mix of categorical and numeric variables.
-
 ## OSAT
 
 OSAT provides templates for plates that hold 2, 4, 8 Illumina BeadChip chips, having 24, 48 or 96 wells.
 Moreover, it works for both numeric and categorical variables:
 
-```{r osat, fig.cap="Sample distributions with OSAT."}
+```{r osat, fig.cap="Sample, distributions with OSAT.", include = FALSE}
 library("ggplot2")
 theme_plots <- theme_minimal() + theme(text = element_text(size = 10))
 theme_set(theme_plots)
@@ -135,18 +136,15 @@ osat_smoke <- ggplot(osat_report) +
   geom_bar(aes(Smoke, fill = as.character(plates))) +
   facet_wrap(~plates) +
   labs(fill = "Batch")
-(osat_smoke/(osat_age + osat_sex + plot_layout(guides = 'collect'))) +
-  plot_annotation(title = "OSAT")
 ```
 
-OSAT returned one row less than the input provided.
-This row was deleted because it had an NA value for the Sex variable.
+OSAT might return less rows than the input provided because they might have NA value.
 
 ## anticlust
 
-anticlust does not handle all the variables, it only accepts numeric variables:
+anticlust does not handle all the variables, it only accepts numeric variables.
 
-```{r anticlust, error=TRUE, fig.cap="Sample distributions with anticlust."}
+```{r anticlust, fig.cap="Sample, distributions with anticlust.", include = FALSE}
 library("anticlust")
 anticlust_index <- anticlustering(
   survey[, VoI],
@@ -165,18 +163,18 @@ anticlust_index <- anticlustering(
 report_anticlust <- data.frame(Age = survey[, "Age"], 
                                batch = as.character(anticlust_index))
 library("ggplot2")
[…]
