+++ paper.md
@@ -1,405 +1,260 @@
 ---
-title: 'experDesign: stratifying samples into batches with minimal bias'
+affiliations:
+- index: 1
+  name: Centro de Investigación Biomédica en Red, Enfermedades Hepáticas
+    y Digestivas
+- index: 2
+  name: Institut d’Investigacions Biomèdiques August Pi i Sunyer,
+    IDIBAPS
+authors:
+- affiliation: 1, 2
+  name: Lluís Revilla Sancho
+  orcid: 0000-0001-9747-2570
+- affiliation: 1
+  name: Juan-José Lozano
+  orcid: 0000-0001-9747-2570
+- affiliation: 2
+  name: Azucena Salas
+  orcid: 0000-0001-9747-2570
+bibliography: paper.bib
+date: 25th August 2021
 output:
-  word_document:
-    keep_md: yes
+  html_document:
+    df_print: paged
   md_document:
     preserve_yaml: yes
   pdf_document: default
-  html_document:
-    df_print: paged
-date: "25th August 2021"
-affiliations:
-- name: Centro de Investigación Biomédica en Red, Enfermedades Hepáticas y Digestivas
-  index: 1
-- name: Institut d'Investigacions Biomèdiques August Pi i Sunyer, IDIBAPS
-  index: 2
-authors:
-- name: Lluís Revilla Sancho
-  orcid: 0000-0001-9747-2570
-  affiliation: 1, 2
-- name: Juan-José Lozano
-  orcid: 0000-0001-9747-2570
-  affiliation: '1'
-- name: Azucena Salas
-  orcid: 0000-0001-9747-2570
-  affiliation: '2'
+  word_document:
+    keep_md: yes
 tags:
 - R
 - batch effect
 - experiment design
-bibliography: paper.bib
+title: "experDesign: stratifying samples into batches with minimal bias"
 ---
 
-
+    knitr::opts_chunk$set(collapse = TRUE, warning = TRUE, 
+                          error = 0, include = FALSE,
+                          fig.width = 9)
 
 # Summary
 
-The design of an experiment is critical to its success.
-Nonetheless, even when correctly designed, the process leading up to the moment of measuring a given variable is critical.
-At any one of the several steps, from sample collection to measurement of a variable, various errors and problems can affect the experimental results.
-Failure to take such variability into account can render an experiment inconclusive.
-*experDesign* provides tools to minimize the risk of inconclusive results by assigning samples to batches to reduce potential batch effects.
+The design of an experiment is critical to its success. Nonetheless,
+even when correctly designed, the process leading up to the moment of
+measuring a given variable is critical. At any one of the several steps,
+from sample collection to measurement of a variable, various errors and
+problems can affect the experimental results. Failure to take such
+variability into account can render an experiment inconclusive.
+*experDesign* provides tools to minimize the risk of inconclusive
+results by assigning samples to batches to reduce potential batch
+effects.
 
 # Introduction
 
-To design an experiment that can support conclusive results upon analysis, the source of the variation between samples must be identified.
-Typically, one can control the environment in which the study or experiment is being conducted.
-Sometimes, however, this is not possible.
-In such cases, techniques to control variations must to be applied.
-There are three methods used to decrease the uncertainty of the unwanted variation: blocking, randomization and replication [@klaus2015].
+To design an experiment that can support conclusive results upon
+analysis, the source of the variation between samples must be
+identified. Typically, one can control the environment in which the
+study or experiment is being conducted. Sometimes, however, this is not
+possible. In such cases, techniques to control variations must to be
+applied. There are three methods used to decrease the uncertainty of the
+unwanted variation: blocking, randomization and replication (Klaus
+2015).
 
[…]
