### 2021-11-10 e84cf82
Update plots and text accordingly
