### 2021-11-10 6b3ee87
Fix legend and add a sentence about examples
