### 2021-11-10 f7cc759
Fix caption and mention the examples on the help files
