### 2021-11-10 #3358 @llrs
Sorry for all the notifications, I think that now it is ready for publication.
@lpantano I removed all the code and left only two plots and the scheme, the page length is still 5 pages though.
