### 2021-11-18 #3358 @lpantano
Thank you, I think we can do a final edit. I would remove figure 1 and add the text under each tool section together with the bullet points you have there. I think that would leave a good structure to the paper, with enough information for the reader to go to other tools. You can leave a description in the text what is the difference with figure 1 without having the figure. Let me know if this makes sense to you, and after that I will accept it, whatever is the number of pages. Thanks!
