### 2021-11-18 a853e97
Remove figure 1, move text and thank editor and reviewers.
