+++ paper.md
@@ -22,10 +22,10 @@ output:
   html_document:
     df_print: paged
   md_document:
-    preserve_yaml: yes
+    preserve_yaml: true
   pdf_document: default
   word_document:
-    keep_md: yes
+    keep_md: true
 tags:
 - R
 - batch effect
@@ -114,52 +114,34 @@ packages:
     the other variables (Yan et al. 2012). This algorithm relies on
     categorical variables and cannot use numerical variables (e.g.,
     those that are age- or time-related) unless they are treated as
-    categorical variables.
+    categorical variables.  
+    OSAT provides templates for plates that hold 2, 4, 8 Illumina
+    BeadChip chips, having 24, 48 or 96 wells. Moreover, it works for
+    both numeric and categorical variables but OSAT might return less
+    rows than the input provided because they might have NA value.
 
 -   [*anticlust*](https://cran.r-project.org/package=anticlust), at
     CRAN, divides the samples into similar groups, ensuring similarity
     by enforcing heterogeneity within groups (Papenberg and Klau 2020).
-    Conceptually it is similar to the clustering method k-means.
+    Conceptually it is similar to the clustering method k-means.  
+    anticlust does not handle all types of variables, it only accepts
+    numeric variables.
 
 -   Recently, [*Omixer*](https://bioconductor.org/packages/Omixer/), a
     new package, has been made available at Bioconductor (Sinke, Cats,
     and Heijmans 2021). It tests whether the random assignments made by
     it are homogeneous by transforming all variables to numeric values
     and using the Kendall’s correlation when there are more than 5
-    samples; otherwise, it utilizes the Pearson’s chi-squared test.
+    samples; otherwise, it utilizes the Pearson’s chi-squared test.  
+    There is a bug in the Omixer that prevents it from working unless
+    specific conditions are meet. This precluded any comparisons of
+    Omixer with other tools using the same settings.
 
 For completeness a description and comparison of the usage of the
 different software packages currently available on CRAN and Bioconductor
 is presented below. First, we start with some real data obtained from a
 survey. This data set has three variables of interest; Sex, Smoke and
 Age are a mix of categorical and numeric variables.
-
-## OSAT
-
-OSAT provides templates for plates that hold 2, 4, 8 Illumina BeadChip
-chips, having 24, 48 or 96 wells. Moreover, it works for both numeric
-and categorical variables but OSAT might return less rows than the input
-provided because they might have NA value.
-
-## anticlust
-
-anticlust does not handle all types of variables, it only accepts
-numeric variables.
-
-## Omixer
-
-There is a bug in the Omixer that prevents it from working unless
-specific conditions are meet. This precluded any comparisons of Omixer
-with other tools using the same settings.
-
-These methods results on this distributions:
-
-![Figure A: distribution of samples according to *OSAT*, Figure B
-according to *anticlust* on the variables used (1: Smoke, 2: Age, 3:
-Sex).](paper_files/figure-markdown_strict/panel1-1.png)
-
-On Figure 1, we can see that the distributions is fairly similar between
-these two packages.
 
 # Statement of need
 
@@ -201,8 +183,12 @@ each batch, followed by the `spatial` function to randomly distribute
 the samples homogeneously by position within each batch. See the example
 in `inspect` and the vignette:
 
-![Sample, distributions with
+![Example of distributions generated with
 ***experDesign***.](paper_files/figure-markdown_strict/experDesign-1.png)
+
+On Figure 1 one can see that the distribution index generated by
+experDesign uses both numeric and categorical variables keeping also the
+samples with missing (NA) information.
 
 The statistics of the index can be checked for multiple statistics, as
[…]
