+++ paper.Rmd
@@ -88,13 +88,18 @@ There are certain tools that can minimize batch effects on the R language in mul
 Here we briefly describe the currently available packages:
 
 -   [*OSAT*](https://bioconductor.org/packages/OSAT/), at Bioconductor, first allocates the samples from each batch according to a variable; it then shuffles the samples from each batch in order to randomize the other variables [@yan2012].
-    This algorithm relies on categorical variables and cannot use numerical variables (e.g., those that are age- or time-related) unless they are treated as categorical variables.
+    This algorithm relies on categorical variables and cannot use numerical variables (e.g., those that are age- or time-related) unless they are treated as categorical variables.\
+    OSAT provides templates for plates that hold 2, 4, 8 Illumina BeadChip chips, having 24, 48 or 96 wells.
+    Moreover, it works for both numeric and categorical variables but OSAT might return less rows than the input provided because they might have NA value.
 
 -   [*anticlust*](https://cran.r-project.org/package=anticlust), at CRAN, divides the samples into similar groups, ensuring similarity by enforcing heterogeneity within groups [@papenberg2020].
-    Conceptually it is similar to the clustering method k-means.
+    Conceptually it is similar to the clustering method k-means.\
+    anticlust does not handle all types of variables, it only accepts numeric variables.
 
 -   Recently, [*Omixer*](https://bioconductor.org/packages/Omixer/), a new package, has been made available at Bioconductor [@sinke2021].
-    It tests whether the random assignments made by it are homogeneous by transforming all variables to numeric values and using the Kendall's correlation when there are more than 5 samples; otherwise, it utilizes the Pearson's chi-squared test.
+    It tests whether the random assignments made by it are homogeneous by transforming all variables to numeric values and using the Kendall's correlation when there are more than 5 samples; otherwise, it utilizes the Pearson's chi-squared test.\
+    There is a bug in the Omixer that prevents it from working unless specific conditions are meet.
+    This precluded any comparisons of Omixer with other tools using the same settings.
 
 For completeness a description and comparison of the usage of the different software packages currently available on CRAN and Bioconductor is presented below.
 First, we start with some real data obtained from a survey.
@@ -108,11 +113,6 @@ n_batch <- 3
 size_subset <- 96
 iterations <- 1000
 ```
-
-## OSAT
-
-OSAT provides templates for plates that hold 2, 4, 8 Illumina BeadChip chips, having 24, 48 or 96 wells.
-Moreover, it works for both numeric and categorical variables but OSAT might return less rows than the input provided because they might have NA value.
 
 ```{r osat, fig.cap="Sample, distributions with OSAT.", include = FALSE}
 library("ggplot2")
@@ -141,12 +141,6 @@ osat_smoke <- ggplot(osat_report) +
   labs(fill = "Batch")
 ```
 
-
-
-## anticlust
-
-anticlust does not handle all types of variables, it only accepts numeric variables.
-
 ```{r anticlust, fig.cap="Sample, distributions with anticlust.", include = FALSE}
 library("anticlust")
 anticlust_index <- anticlustering(
@@ -172,11 +166,6 @@ anticlust <- ggplot(report_anticlust) +
   labs(fill = "Batch")
 ```
 
-## Omixer
-
-There is a bug in the Omixer that prevents it from working unless specific conditions are meet.
-This precluded any comparisons of Omixer with other tools using the same settings.
-
 ```{r Omixer, include = FALSE}
 library("Omixer")
 survey$ID <- seq_len(nrow(survey))
[…]
