+++ paper.Rmd
@@ -195,7 +195,7 @@ Omixer_index <- omixerRand(survey, sampleId = "ID",
 
 Current solutions for stratifying samples to reduce and control batch effect do not work for all cases.
 They are either specialized to a particular type of data, they omit some conditions that are usually met, or they only work under a specific subset of conditions.
-The new package ***experDesign*** works with all data types and does not require a spatial distribution making it suitable for all kind of experiments.
+The new package [***experDesign***](https://cran.r-project.org/package=experDesign "experDesign on CRAN") works with all data types and does not require a spatial distribution making it suitable for all kind of experiments.
 This package is intended for people needing a quick and easy solution that will provide reasonable suggestions on how to best distribute the samples for analysis.
 
 # Description
@@ -304,7 +304,7 @@ optimum_batches(size_data = 250, size_subset = 96)
 sizes_batches(size_data = 250, size_subset = 96, batches = 3)
 ```
 
-In conclusion ***experDesign*** offers a fast method for preparing a batched experiment.
+In conclusion [***experDesign***](https://cran.r-project.org/package=experDesign "experDesign pacakge") offers a fast method for preparing a batched experiment.
 It can use as many numeric and categorical variables as needed to stratify the experimental design based on batches including spatial distributions.
 
 # Acknowledgments
