+++ README.md
@@ -16,6 +16,8 @@ stable](https://img.shields.io/badge/lifecycle-stable-brightgreen.svg)](https://
 [![Project Status: Active - The project has reached a stable, usable
 state and is being actively
 developed.](https://www.repostatus.org/badges/latest/active.svg)](https://www.repostatus.org/#active)
+[![DOI](https://zenodo.org/badge/142569201.svg)](https://zenodo.org/badge/latestdoi/142569201)
+
 <!-- badges: end -->
 
 The goal of experDesign is to help you decide which samples go in which
@@ -91,32 +93,33 @@ head(survey[, keep])
 #> 4   Male Right Never 20.333
 #> 5   Male Right Never 23.667
 #> 6 Female Right Never 21.000
+
 # Looking for groups at most of 70 samples.
 index <- design(pheno = survey, size_subset = 70, omit = omit)
 index
 #> $SubSet1
-#>  [1]   1   2   4   6  11  18  22  25  38  40  44  46  48  50  55  60  63  67  70
-#> [20]  85  91  92  99 100 103 105 110 116 120 126 129 134 141 144 145 149 160 165
-#> [39] 168 171 172 174 178 185 190 196 197 198 201 204 206 208 209 210 213 216 218
-#> [58] 228 231 232
+#>  [1]   8  10  11  13  16  18  21  25  26  32  38  42  43  44  50  54  55  60  65
+#> [20]  67  68  72  74  75  76  79  88  89  92  97 109 124 140 143 145 147 150 155
+#> [39] 163 164 173 176 181 182 183 185 187 194 200 206 207 208 209 210 213 217 218
+#> [58] 228 231 236
 #> 
 #> $SubSet2
-#>  [1]   7   9  12  13  15  17  23  30  43  47  49  58  62  64  66  69  71  73  76
-#> [20]  79  80  81  88  90  98 102 108 109 114 115 118 121 123 124 125 128 131 132
-#> [39] 139 146 148 151 153 155 159 164 169 175 177 181 182 183 194 211 217 222 225
-#> [58] 227 233
+#>  [1]   2   3   7   9  12  15  20  23  28  36  37  39  45  49  51  52  57  59  61
+#> [20]  63  64  66  69  70  71  82  84  85  86  99 100 104 112 114 115 119 126 135
+#> [39] 137 146 153 159 160 166 169 171 174 175 180 186 191 195 204 211 215 219 221
+#> [58] 227 232
 #> 
 #> $SubSet3
-#>  [1]   5   8  14  16  19  20  24  26  27  28  29  34  35  37  41  42  51  53  61
-#> [20]  72  74  75  78  83  87  89  95  96 104 107 113 117 119 122 133 140 143 150
-#> [39] 152 154 156 157 161 166 179 184 186 187 188 191 193 195 202 207 214 224 226
-#> [58] 230 237
+#>  [1]   4   6  14  24  29  34  35  40  46  47  53  56  58  73  78  80  81  87  95
+#> [20]  96  98 101 102 103 105 107 108 117 118 120 121 122 125 129 131 132 133 134
+#> [39] 136 139 142 148 152 154 156 157 162 172 189 190 199 205 220 226 229 230 233
+#> [58] 234 235
 #> 
 #> $SubSet4
-#>  [1]   3  10  21  31  32  33  36  39  45  52  54  56  57  59  65  68  77  82  84
-#> [20]  86  93  94  97 101 106 111 112 127 130 135 136 137 138 142 147 158 162 163
-#> [39] 167 170 173 176 180 189 192 199 200 203 205 212 215 219 220 221 223 229 234
-#> [58] 235 236
[…]
