### 2021-11-22 #3358 @llrs
After merging the paper branch with the main branch of the software I made a tagged release with version: [0.1.1](https://github.com/llrs/experDesign/releases/tag/v0.1.1).
I updated archive on Zenodo at: https://doi.org/10.5281/zenodo.5718051 (I copied the title and author's data copied from the manuscript).
I haven't submitted the new version of the package to CRAN, let me know if this is also needed.
