+++ paper.md
@@ -174,9 +174,8 @@ the best sample distribution is returned to the user.
 Users can examine the following flowchart to decide what function(s)
 they need to use:
 
-<figure>
-<img src="flow_chart.png" width="353" alt="Flow chart to decide which functions are needed" /><figcaption aria-hidden="true">Flow chart to decide which functions are needed</figcaption>
-</figure>
+
+![Flow chart to decide which functions are needed](flow_chart.png)
 
 If users want a design without replicates but the batches have some
 spatial distribution, we must use `design` to allocate the samples on
@@ -187,7 +186,7 @@ in `inspect` and the vignette:
 ![Example of distributions generated with
 ***experDesign***.](paper_files/figure-markdown_strict/experDesign-1.png)
 
-On Figure 1 one can see that the distribution index generated by
+On Figure 2 one can see that the distribution index generated by
 experDesign uses both numeric and categorical variables keeping also the
 samples with missing (NA) information.
 
