### 2021-11-26 #3358 @llrs
Hi, many thanks for accepting the manuscript. I noticed that when merging the branches the flow chart has disappeared from the pdf. Not sure how to proceed as the [preview shows](https://github.com/llrs/experDesign/blob/master/paper.md) the figure and I haven't modified that (till latest commit after releasing 0.1.1)
