### 2021-11-27 #3358 @Kevin-Mattheus-Moerman
@llrs @lpantano I have proofread the paper and have inspected the archived software on ZENODO and found all to be in order. I will now proceed to accept this work in JOSS.
